requirements.txt
ciscoconfparse==1.9.39
fastapi==0.104.1
fastapi_utils==0.2.1
netmiko==4.3.0
pydantic==1.10.13
pymongo==4.6.0
pyserial==3.5
Requests==2.31.0
uvicorn==0.24.0.post1
